# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login pengguna
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Ambil data pengguna saat ini
      responses:
        "200":
          description: Data pengguna
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Tidak terautentikasi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # DASHBOARD
...
  /guru:
    get:
      operationId: listGuru
      tags: [guru]
      summary: Daftar semua guru
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daftar guru
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guru"
    post:
      operationId: createGuru
      tags: [guru]
      summary: Tambah guru baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuruBody"
      responses:
        "201":
          description: Guru berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
  /guru/{id}:
    get:
      operationId: getGuru
      tags: [guru]
      summary: Detail guru
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Data guru
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
    put:
      operationId: updateGuru
      tags: [guru]
      summary: Update data guru
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuruBody"
      responses:
        "200":
          description: Data guru diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guru"
    delete:
      operationId: deleteGuru
      tags: [guru]
      summary: Hapus guru
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guru berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # KELAS
...
  /jadwal:
    get:
      operationId: listJadwal
      tags: [jadwal]
      summary: Daftar jadwal pelajaran
      parameters:
        - name: kelas_id
          in: query
          schema:
            type: string
        - name: guru_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daftar jadwal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Jadwal"
    post:
      operationId: createJadwal
      tags: [jadwal]
      summary: Tambah jadwal baru
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJadwalBody"
      responses:
        "201":
          description: Jadwal berhasil ditambahkan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Jadwal"
  /jadwal/{id}:
    delete:
      operationId: deleteJadwal
      tags: [jadwal]
      summary: Hapus jadwal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Jadwal berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ABSENSI
  /absensi:
    get:
      operationId: listAbsensi
      tags: [absensi]
      summary: Daftar absensi
      parameters:
        - name: siswa_id
          in: query
          schema:
            type: string
        - name: kelas_id
          in: query
          schema:
            type: string
        - name: mata_pelajaran_id
          in: query
          schema:
            type: string
        - name: tanggal
          in: query
          schema:
            type: string
        - name: bulan
          in: query
          schema:
            type: string
            description: Format YYYY-MM
      responses:
        "200":
          description: Daftar absensi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Absensi"
    post:
      operationId: createAbsensi
      tags: [absensi]
      summary: Input absensi (bisa bulk)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAbsensiBody"
      responses:
        "201":
          description: Absensi berhasil disimpan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Absensi"
  /absensi/rekap:
    get:
      operationId: getRekapAbsensi
      tags: [absensi]
      summary: Rekap absensi per siswa per bulan
      parameters:
        - name: siswa_id
          in: query
          schema:
            type: string
        - name: bulan
          in: query
          schema:
            type: string
            description: Format YYYY-MM
        - name: kelas_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Rekap absensi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RekapAbsensi"
  /absensi/{id}:
    put:
      operationId: updateAbsensi
      tags: [absensi]
      summary: Update status absensi
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAbsensiBody"
      responses:
        "200":
          description: Absensi diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Absensi"

  # NILAI
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        access_token:
          type: string
        refresh_token:
          type: string
      required:
        - user
        - access_token
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, guru, siswa]
        full_name:
          type: string
        created_at:
          type: string
      required:
        - id
        - email
        - role
        - full_name
...
    Siswa:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
          nullable: true
        nis:
          type: string
        nisn:
          type: string
        nama:
          type: string
        kelas_id:
          type: string
          nullable: true
        kelas:
          $ref: "#/components/schemas/Kelas"
          nullable: true
        jenis_kelamin:
          type: string
          enum: [L, P]
        tanggal_lahir:
          type: string
        alamat:
          type: string
          nullable: true
        no_hp_ortu:
          type: string
          nullable: true
        foto_url:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - nis
        - nisn
        - nama
        - jenis_kelamin
        - tanggal_lahir
...
    Guru:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
          nullable: true
        nip:
          type: string
          nullable: true
        nama:
          type: string
        mata_pelajaran_id:
          type: string
          nullable: true
        mata_pelajaran:
          $ref: "#/components/schemas/MataPelajaran"
          nullable: true
        no_hp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        created_at:
          type: string
      required:
        - id
        - nama
    CreateGuruBody:
      type: object
      properties:
        nip:
          type: string
          nullable: true
        nama:
          type: string
        mata_pelajaran_id:
          type: string
          nullable: true
        no_hp:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      required:
        - nama
    UpdateGuruBody:
      type: object
      properties:
        nip:
          type: string
          nullable: true
        nama:
          type: string
        mata_pelajaran_id:
          type: string
          nullable: true
        no_hp:
          type: string
          nullable: true
    Kelas:
      type: object
      properties:
        id:
          type: string
        nama_kelas:
          type: string
        tingkat:
          type: integer
        wali_kelas_id:
          type: string
          nullable: true
        wali_kelas:
          $ref: "#/components/schemas/Guru"
          nullable: true
        tahun_ajaran:
          type: string
        jumlah_siswa:
          type: integer
          nullable: true
      required:
        - id
        - nama_kelas
        - tingkat
        - tahun_ajaran
...
    MataPelajaran:
      type: object
      properties:
        id:
          type: string
        nama_mapel:
          type: string
        kode_mapel:
          type: string
          nullable: true
        guru_id:
          type: string
          nullable: true
        guru:
          $ref: "#/components/schemas/Guru"
          nullable: true
      required:
        - id
        - nama_mapel
...
    Jadwal:
      type: object
      properties:
        id:
          type: string
        kelas_id:
          type: string
        kelas:
          $ref: "#/components/schemas/Kelas"
          nullable: true
        mata_pelajaran_id:
          type: string
        mata_pelajaran:
          $ref: "#/components/schemas/MataPelajaran"
          nullable: true
        guru_id:
          type: string
        guru:
          $ref: "#/components/schemas/Guru"
          nullable: true
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jam_mulai:
          type: string
        jam_selesai:
          type: string
      required:
        - id
        - kelas_id
        - mata_pelajaran_id
        - guru_id
        - hari
        - jam_mulai
        - jam_selesai
    CreateJadwalBody:
      type: object
      properties:
        kelas_id:
          type: string
        mata_pelajaran_id:
          type: string
        guru_id:
          type: string
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jam_mulai:
          type: string
        jam_selesai:
          type: string
      required:
        - kelas_id
        - mata_pelajaran_id
        - guru_id
        - hari
        - jam_mulai
        - jam_selesai
    Absensi:
      type: object
      properties:
        id:
          type: string
        siswa_id:
          type: string
        siswa:
          $ref: "#/components/schemas/Siswa"
          nullable: true
        mata_pelajaran_id:
          type: string
        mata_pelajaran:
          $ref: "#/components/schemas/MataPelajaran"
          nullable: true
        tanggal:
          type: string
        status:
          type: string
          enum: [hadir, izin, sakit, alfa]
        keterangan:
          type: string
          nullable: true
      required:
        - id
        - siswa_id
        - mata_pelajaran_id
        - tanggal
        - status
    CreateAbsensiBody:
      type: object
      properties:
        records:
          type: array
          items:
            type: object
            properties:
              siswa_id:
                type: string
              mata_pelajaran_id:
                type: string
              tanggal:
                type: string
              status:
                type: string
                enum: [hadir, izin, sakit, alfa]
              keterangan:
                type: string
                nullable: true
            required:
              - siswa_id
              - mata_pelajaran_id
              - tanggal
              - status
      required:
        - records
    UpdateAbsensiBody:
      type: object
      properties:
        status:
          type: string
          enum: [hadir, izin, sakit, alfa]
        keterangan:
          type: string
          nullable: true
      required:
        - status
    RekapAbsensiItem:
      type: object
      properties:
        hadir:
          type: integer
        izin:
          type: integer
        sakit:
          type: integer
        alfa:
          type: integer
        total:
          type: integer
      required:
        - hadir
        - izin
        - sakit
        - alfa
        - total
    RekapAbsensi:
      type: object
      properties:
        bulan:
          type: string
        siswa:
          type: array
          items:
            type: object
            properties:
              siswa_id:
                type: string
              nama:
                type: string
              nis:
                type: string
              rekap:
                $ref: "#/components/schemas/RekapAbsensiItem"
            required:
              - siswa_id
              - nama
              - nis
              - rekap
      required:
        - bulan
        - siswa
...