# lib/api-zod/src/generated/types/createGuruBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGuruBody {
  nip?: string | null;
  nama: string;
  mata_pelajaran_id?: string | null;
  no_hp?: string | null;
  email?: string | null;
  password?: string | null;
}