# lib/api-zod/src/generated/types/guru.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */
import type { MataPelajaran } from "./mataPelajaran";

export interface Guru {
  id: string;
  user_id?: string | null;
  nip?: string | null;
  nama: string;
  mata_pelajaran_id?: string | null;
  mata_pelajaran?: MataPelajaran | null;
  no_hp?: string | null;
  email?: string | null;
  created_at?: string;
}