# lib/api-zod/src/generated/types/jadwal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */
import type { Guru } from "./guru";
import type { JadwalHari } from "./jadwalHari";
import type { Kelas } from "./kelas";
import type { MataPelajaran } from "./mataPelajaran";

export interface Jadwal {
  id: string;
  kelas_id: string;
  kelas?: Kelas | null;
  mata_pelajaran_id: string;
  mata_pelajaran?: MataPelajaran | null;
  guru_id: string;
  guru?: Guru | null;
  hari: JadwalHari;
  jam_mulai: string;
  jam_selesai: string;
}