# lib/api-zod/src/generated/types/kelas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */
import type { Guru } from "./guru";

export interface Kelas {
  id: string;
  nama_kelas: string;
  tingkat: number;
  wali_kelas_id?: string | null;
  wali_kelas?: Guru | null;
  tahun_ajaran: string;
  jumlah_siswa?: number | null;
}