# lib/api-zod/src/generated/types/rekapAbsensiItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */

export interface RekapAbsensiItem {
  hadir: number;
  izin: number;
  sakit: number;
  alfa: number;
  total: number;
}