# lib/api-zod/src/generated/types/siswa.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SIAKAD - Sistem Informasi Akademik SMP Negeri 2 Rambang
 * OpenAPI spec version: 0.1.0
 */
import type { Kelas } from "./kelas";
import type { SiswaJenisKelamin } from "./siswaJenisKelamin";

export interface Siswa {
  id: string;
  user_id?: string | null;
  nis: string;
  nisn: string;
  nama: string;
  kelas_id?: string | null;
  kelas?: Kelas | null;
  jenis_kelamin: SiswaJenisKelamin;
  tanggal_lahir: string;
  alamat?: string | null;
  no_hp_ortu?: string | null;
  foto_url?: string | null;
  created_at?: string;
}